Move classes from org.embulk.cli.parse into org.embulk.deps.cli
