+++ embulk-core/src/main/java/org/embulk/deps/cli/EmbulkCommandLineHelpRequired.java
@@ -1,4 +1,4 @@
-package org.embulk.cli.parse;
+package org.embulk.deps.cli;
 
 public final class EmbulkCommandLineHelpRequired extends Throwable {
     EmbulkCommandLineHelpRequired() {
