+++ embulk-core/src/main/java/org/embulk/deps/cli/EmbulkCommandLineParseException.java
@@ -1,4 +1,4 @@
-package org.embulk.cli.parse;
+package org.embulk.deps.cli;
 
 public final class EmbulkCommandLineParseException extends Exception {
     protected EmbulkCommandLineParseException() {
