+++ embulk-core/src/main/java/org/embulk/deps/cli/HelpMessageAsCliOption.java
@@ -1,4 +1,4 @@
-package org.embulk.cli.parse;
+package org.embulk.deps.cli;
 
 import org.apache.commons.cli.Option;
