+++ embulk-core/src/main/java/org/embulk/deps/cli/OptionDefinition.java
@@ -1,4 +1,4 @@
-package org.embulk.cli.parse;
+package org.embulk.deps.cli;
 
 import org.apache.commons.cli.Option;
 import org.embulk.cli.EmbulkCommandLine;
